Move AgentFrameworkSpanProcessor + helpers to AgentFramework/ folder
Relocate processor, helper, and model from Agent365/Extensions/AgentFramework/
to AgentFramework/ to match their registration point (UseAgentFramework).
Update namespaces from Microsoft.OpenTelemetry.Agent365.Extensions.AgentFramework
to Microsoft.OpenTelemetry.AgentFramework.
